spikes/spike5_celery_redis/requirements.txt
celery[redis]>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-timeout>=2.3.0
